refactoring to create separate executors for Recipe, Project and Swagger specs
