+++ src/main/java/com/smartbear/readyapi/client/execution/RecipeExecutor.java
@@ -1,86 +1,24 @@
 package com.smartbear.readyapi.client.execution;
 
-import com.smartbear.readyapi.client.ExecutionListener;
 import com.smartbear.readyapi.client.RecipeFilter;
-import com.smartbear.readyapi.client.RepositoryProjectExecutionRequest;
 import com.smartbear.readyapi.client.TestRecipe;
-import com.smartbear.readyapi.client.model.HarLogRoot;
-import com.smartbear.readyapi.client.model.ProjectResultReport;
-import com.smartbear.readyapi.client.model.ProjectResultReports;
-import com.smartbear.readyapi.client.model.RequestTestStepBase;
 import com.smartbear.readyapi.client.model.TestCase;
-import com.smartbear.readyapi.client.model.TestStep;
-import com.smartbear.readyapi.client.model.UnresolvedFile;
-import io.swagger.client.auth.HttpBasicAuth;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import javax.annotation.Nullable;
-import java.io.File;
-import java.net.URL;
-import java.util.ArrayList;
 import java.util.List;
-import java.util.Timer;
-import java.util.TimerTask;
 import java.util.concurrent.CopyOnWriteArrayList;
 
 /**
  * Responsible for executing test recipes and projects on a Ready! API Server, synchronously or asynchronously.
  */
-public class RecipeExecutor {
+public class RecipeExecutor extends AbstractExecutor {
+    private static Logger logger = LoggerFactory.getLogger(RecipeExecutor.class);
 
-    private static Logger logger;
-
-    static {
-        if (System.getProperty("org.slf4j.simpleLogger.defaultLogLevel") == null) { //Don't set if user has defined the log level
-            System.setProperty("org.slf4j.simpleLogger.defaultLogLevel", "trace");
-        }
-        logger = LoggerFactory.getLogger(RecipeExecutor.class);
-    }
-
-    enum SwaggerFormat {
-        JSON("application/json"),
-        YAML("application/yaml");
-
-        private final String mimeType;
-
-        SwaggerFormat(String mimeType) {
-            this.mimeType = mimeType;
-        }
-
-        public String getMimeType() {
-            return mimeType;
-        }
-    }
-
-    private static final int NUMBER_OF_RETRIES_IN_CASE_OF_ERRORS = 3;
-    private final TestServerApi apiStub;
-    private HttpBasicAuth authentication;
-    private final List<ExecutionListener> executionListeners = new CopyOnWriteArrayList<>();
     private final List<RecipeFilter> recipeFilters = new CopyOnWriteArrayList<>();
 
-    public RecipeExecutor(Scheme scheme, String host, int port) {
-        this(scheme, host, port, ServerDefaults.VERSION_PREFIX, new CodegenBasedTestServerApi());
-    }
-
-    public RecipeExecutor(String host, int port) {
-        this(ServerDefaults.DEFAULT_SCHEME, host, port);
-    }
-
-    public RecipeExecutor(String host) {
-        this(host, ServerDefaults.DEFAULT_PORT);
-    }
-
-    // Used for testing
-    RecipeExecutor(Scheme scheme, String host, int port, String basePath, TestServerApi apiStub) {
-        this.apiStub = apiStub;
-        apiStub.setBasePath(String.format("%s://%s:%d%s", scheme.getValue(), host, port, basePath));
-    }
-
-    public void setCredentials(String username, String password) {
-        authentication = new HttpBasicAuth();
-        authentication.setUsername(username);
-        authentication.setPassword(password);
+    RecipeExecutor(TestServerClient testServerClient) {
+        super(testServerClient);
     }
 
     public void addRecipeFilter(RecipeFilter recipeFilter) {
@@ -91,90 +29,7 @@ public class RecipeExecutor {
         recipeFilters.remove(recipeFilter);
     }
 
-    public void addExecutionListener(ExecutionListener listener) {
-        executionListeners.add(listener);
-    }
-
-    public void removeExecutionListener(ExecutionListener listener) {
-        executionListeners.remove(listener);
-    }
-
-    /**
-     * @deprecated Use RecipeExecutor#submitProject(ProjectExecutionRequest) instead.
-     */
-    @Deprecated
-    public Execution submitProject(File project) {
-        return submitProject(project, null, null, null);
-    }
-
-    public Execution submitRepositoryProject(RepositoryProjectExecutionRequest executionRequest) {
-        Execution execution = doExecuteProjectFromRepository(executionRequest, true);
-        notifyRequestSubmitted(execution);
-        return execution;
-    }
-
-    public Execution executeRepositoryProject(RepositoryProjectExecutionRequest executionRequest) {
-        Execution execution = doExecuteProjectFromRepository(executionRequest, false);
-        if (execution != null) {
-            notifyExecutionFinished(execution.getCurrentReport());
-        }
-        return execution;
-    }
-
-    /**
-     * @deprecated Use RecipeExecutor#submitProject(ProjectExecutionRequest) instead.
-     */
-    @Deprecated
-    public Execution submitProject(File project, @Nullable String testCaseName, @Nullable String testSuiteName,
-                                   @Nullable String environment) throws ApiException {
-        ProjectExecutionRequest executionRequest = ProjectExecutionRequest.Builder.newInstance()
-                .withProjectFile(project)
-                .forTestCase(testCaseName)
-                .forTestSuite(testSuiteName)
-                .forEnvironment(environment)
-                .build();
-        return submitProject(executionRequest);
-    }
-
-    public Execution submitProject(ProjectExecutionRequest projectExecutionRequest) throws ApiException {
-        Execution execution = doExecuteProject(projectExecutionRequest, true);
-        notifyRequestSubmitted(execution);
-        return execution;
-    }
-
-    /**
-     * @deprecated Use RecipeExecutor#executeProject(ProjectExecutionRequest) instead.
-     */
-    @Deprecated
-    public Execution executeProject(File project) {
-        return executeProject(project, null, null, null);
-    }
-
-    /**
-     * @deprecated Use RecipeExecutor#executeProject(ProjectExecutionRequest) instead.
-     */
-    @Deprecated
-    public Execution executeProject(File project, @Nullable String testCaseName, @Nullable String testSuiteName,
-                                    @Nullable String environment) throws ApiException {
-        ProjectExecutionRequest executionRequest = ProjectExecutionRequest.Builder.newInstance()
-                .withProjectFile(project)
-                .forTestCase(testCaseName)
-                .forTestSuite(testSuiteName)
-                .forEnvironment(environment)
-                .build();
-        return executeProject(executionRequest);
-    }
-
-    public Execution executeProject(ProjectExecutionRequest projectExecutionRequest) throws ApiException {
-        Execution execution = doExecuteProject(projectExecutionRequest, false);
-        if (execution != null) {
-            notifyExecutionFinished(execution.getCurrentReport());
-        }
-        return execution;
-    }
-
     public Execution submitRecipe(TestRecipe recipe) throws ApiException {
-
         for (RecipeFilter recipeFilter : recipeFilters) {
             recipeFilter.filterRecipe(recipe);
         }
@@ -192,139 +47,11 @@ public class RecipeExecutor {
         return execution;
     }
 
-    /**
-     * Asynchronous execution of Swagger Definition:
-     * Submit Swagger specification file to TestServer to create and execute tests for each api defined in specifications.
-     *
-     * @param swaggerFile   Swagger file
-     * @param swaggerFormat format
-     * @param endpoint      endpoint against which tests should be executed.
-     *                      Tests will be executed against the host specified in Swagger definition if endpoint is not provided.
-     * @return execution with executionId to be used to query the status
-     */
-    public Execution submitSwagger(File swaggerFile, SwaggerFormat swaggerFormat, String endpoint) {
-        ProjectResultReport projectResultReport = apiStub.postSwagger(swaggerFile, swaggerFormat, endpoint, true, authentication);
-        return new Execution(apiStub, authentication, projectResultReport);
-    }
-
-    /**
-     * Synchronous execution of Swagger Definition:
-     * Submit Swagger specification file to TestServer to create and execute tests for each api defined in specifications.
-     *
-     * @param swaggerFile   Swagger file
-     * @param swaggerFormat format
-     * @param endpoint      endpoint against which tests should be executed.
-     *                      Tests will be executed against the host specified in Swagger definition if endpoint is not provided.
-     * @return execution with execution report
-     */
-    public Execution executeSwagger(File swaggerFile, SwaggerFormat swaggerFormat, String endpoint) {
-        ProjectResultReport projectResultReport = apiStub.postSwagger(swaggerFile, swaggerFormat, endpoint, false, authentication);
-        return new Execution(apiStub, authentication, projectResultReport);
-    }
-
-    /**
-     * Asynchronous : Submit URL of Swagger specification to TestServer to create and execute tests for each api defined in specifications.
-     *
-     * @param swaggerApiURL URL of Swagger API
-     * @param endpoint      endpoint against which tests should be executed.
-     *                      Tests will be executed against the host specified in Swagger definition if endpoint is not provided.
-     * @return execution with executionId to be used to query the status
-     */
-    public Execution submitSwagger(URL swaggerApiURL, String endpoint) {
-        ProjectResultReport projectResultReport = apiStub.postSwagger(swaggerApiURL, endpoint, true, authentication);
-        return new Execution(apiStub, authentication, projectResultReport);
-    }
-
-    /**
-     * Synchronous execution: Submit URL of Swagger specification to TestServer to create and execute tests for each api defined in specifications.
-     *
-     * @param swaggerApiURL URL of Swagger API
-     * @param endpoint      endpoint against which tests should be executed.
-     *                      Tests will be executed against the host specified in Swagger definition if endpoint is not provided.
-     * @return execution with execution report
-     */
-    public Execution executeSwagger(URL swaggerApiURL, String endpoint) {
-        ProjectResultReport projectResultReport = apiStub.postSwagger(swaggerApiURL, endpoint, false, authentication);
-        return new Execution(apiStub, authentication, projectResultReport);
-    }
-
-    public Execution cancelExecution(final Execution execution) {
-        ProjectResultReport projectResultReport = apiStub.cancelExecution(execution.getId(), authentication);
-        execution.addResultReport(projectResultReport);
-        return execution;
-    }
-
-    public HarLogRoot getTransactionLog(final Execution execution, String transactionId) {
-        return apiStub.getTransactionLog(execution.getId(), transactionId, authentication);
-    }
-
-    public List<Execution> getExecutions() {
-        List<Execution> executions = new ArrayList<>();
-        ProjectResultReports projectResultReport = apiStub.getExecutions(authentication);
-        for (ProjectResultReport resultReport : projectResultReport.getProjectResultReports()) {
-            executions.add(new Execution(apiStub, authentication, resultReport));
-        }
-        return executions;
-    }
-
-    private void notifyRequestSubmitted(Execution execution) {
-        if (execution != null) {
-            for (ExecutionListener executionListener : executionListeners) {
-                executionListener.requestSent(execution.getCurrentReport());
-            }
-            new ExecutionStatusChecker(execution).start();
-        }
-    }
-
-    private void notifyErrorOccurred(Exception e) {
-        for (ExecutionListener executionListener : executionListeners) {
-            executionListener.errorOccurred(e);
-        }
-    }
-
-    private void notifyExecutionFinished(ProjectResultReport executionStatus) {
-        for (ExecutionListener executionListener : executionListeners) {
-            executionListener.executionFinished(executionStatus);
-        }
-    }
-
-    private Execution doExecuteProjectFromRepository(RepositoryProjectExecutionRequest executionRequest, boolean async) {
-        try {
-            ProjectResultReport projectResultReport = apiStub.postRepositoryProject(executionRequest, async, authentication);
-            cancelExecutionAndThrowExceptionIfPendingDueToMissingClientCertificate(projectResultReport, null);
-            return new Execution(apiStub, authentication, projectResultReport);
-        } catch (ApiException e) {
-            notifyErrorOccurred(e);
-            logger.debug("An error occurred when sending project to server. Details: " + e.toString());
-            throw e;
-        } catch (Exception e) {
-            notifyErrorOccurred(e);
-            logger.debug("An error occurred when sending project to server", e);
-            throw new ApiException(e);
-        }
-    }
-
-    private Execution doExecuteProject(ProjectExecutionRequest projectExecutionRequest, boolean async) throws ApiException {
-        try {
-            ProjectResultReport projectResultReport = apiStub.postProject(projectExecutionRequest, async, authentication);
-            cancelExecutionAndThrowExceptionIfPendingDueToMissingClientCertificate(projectResultReport, null);
-            return new Execution(apiStub, authentication, projectResultReport);
-        } catch (ApiException e) {
-            notifyErrorOccurred(e);
-            logger.debug("An error occurred when sending project to server. Details: " + e.toString());
-            throw e;
-        } catch (Exception e) {
-            notifyErrorOccurred(e);
-            logger.debug("An error occurred when sending project to server", e);
-            throw new ApiException(e);
-        }
-    }
-
     private Execution doExecuteTestCase(TestCase testCase, boolean async) throws ApiException {
         try {
-            ProjectResultReport projectResultReport = apiStub.postTestRecipe(testCase, async, authentication);
-            cancelExecutionAndThrowExceptionIfPendingDueToMissingClientCertificate(projectResultReport, testCase);
-            return new Execution(apiStub, authentication, projectResultReport);
+            Execution execution = testServerClient.postTestRecipe(testCase, async);
+            cancelExecutionAndThrowExceptionIfPendingDueToMissingClientCertificate(execution.getCurrentReport(), testCase);
+            return execution;
         } catch (ApiException e) {
             notifyErrorOccurred(e);
             logger.debug("An error occurred when sending test recipe to server. Details: " + e.toString());
@@ -333,70 +60,6 @@ public class RecipeExecutor {
             notifyErrorOccurred(e);
             logger.debug("An error occurred when sending test recipe to server", e);
             throw new ApiException(e);
-        }
-    }
-
-    private void cancelExecutionAndThrowExceptionIfPendingDueToMissingClientCertificate(ProjectResultReport projectResultReport, TestCase testCase) {
-        if (ProjectResultReport.StatusEnum.PENDING.equals(projectResultReport.getStatus())) {
-            List<UnresolvedFile> unresolvedFiles = projectResultReport.getUnresolvedFiles();
-            if (unresolvedFiles.size() > 0) {
-                apiStub.cancelExecution(projectResultReport.getExecutionID(), authentication);
-            }
-            for (UnresolvedFile unresolvedFile : unresolvedFiles) {
-                if (testCase == null || unresolvedFile.getFileName().equals(testCase.getClientCertFileName())) {
-                    throw new ApiException(400, "Couldn't find client certificate file: " + unresolvedFile.getFileName());
-                }
-                throwExceptionIfTestStepCertificateIsUnresolved(projectResultReport, testCase, unresolvedFile);
-            }
-        }
-    }
-
-    private void throwExceptionIfTestStepCertificateIsUnresolved(ProjectResultReport projectResultReport, TestCase testCase, UnresolvedFile unresolvedFile) {
-        for (TestStep testStep : testCase.getTestSteps()) {
-            if (testStep instanceof RequestTestStepBase) {
-                RequestTestStepBase requestTestStepBase = (RequestTestStepBase) testStep;
-                if (unresolvedFile.getFileName().equals(requestTestStepBase.getClientCertificateFileName())) {
-                    throw new ApiException(400, "Couldn't find test step client certificate file: " + requestTestStepBase.getClientCertificateFileName());
-                }
-            }
-        }
-    }
-
-    private class ExecutionStatusChecker {
-        private final Timer timer;
-
-        private final Execution execution;
-
-        private int errorCount = 0;
-
-        public ExecutionStatusChecker(Execution execution) {
-            this.execution = execution;
-            timer = new Timer();
-        }
-
-        public void start() {
-            timer.schedule(new CheckingExpireDateTask(), 0, 1000);
-        }
-
-        class CheckingExpireDateTask extends TimerTask {
-            @Override
-            public void run() {
-                try {
-                    ProjectResultReport executionStatus = apiStub.getExecutionStatus(execution.getId(), authentication);
-                    execution.addResultReport(executionStatus);
-                    if (!ProjectResultReport.StatusEnum.RUNNING.equals(executionStatus.getStatus())) {
-                        notifyExecutionFinished(executionStatus);
-                        timer.cancel();
-                    }
-                    errorCount = 0;
-                } catch (Exception e) {
-                    if (errorCount > NUMBER_OF_RETRIES_IN_CASE_OF_ERRORS) {
-                        timer.cancel();
-                    }
-                    logger.debug("Error while checking for execution status", e);
-                    errorCount++;
-                }
-            }
         }
     }
 }
