+++ src/test/java/com/smartbear/readyapi/client/execution/ProjectExecutionTestBase.java
@@ -9,14 +9,20 @@ class ProjectExecutionTestBase {
     static final int PORT = 6234;
     static final String BASE_PATH = "/custom_path";
 
+    TestServerClient testServerClient;
     TestServerApi apiWrapper;
     RecipeExecutor recipeExecutor;
+    ProjectExecutor projectExecutor;
+    SwaggerApiValidator swaggerApiValidator;
 
     @Before
     public void setUpApiMockAndRecipeExecutor() throws Exception {
         apiWrapper = mock(TestServerApi.class);
-        recipeExecutor = new RecipeExecutor(ServerDefaults.DEFAULT_SCHEME, HOST, PORT, BASE_PATH, apiWrapper);
-        recipeExecutor.setCredentials("theUser", "thePassword");
+        testServerClient = new TestServerClient(ServerDefaults.DEFAULT_SCHEME, HOST, PORT, BASE_PATH, apiWrapper);
+        testServerClient.setCredentials("theUser", "thePassword");
+        recipeExecutor = testServerClient.createRecipeExecutor();
+        projectExecutor = testServerClient.createProjectExecutor();
+        swaggerApiValidator = testServerClient.createApiValidator();
     }
 
 }
