+++ src/test/java/com/smartbear/readyapi/client/execution/RecipeExecutorTestCreator.java
@@ -1,7 +1,7 @@
 package com.smartbear.readyapi.client.execution;
 
 public class RecipeExecutorTestCreator {
-    public static RecipeExecutor createRecipeExecutor(Scheme scheme, String host, int port, String basePath, TestServerApi apiStub){
-        return new RecipeExecutor(scheme, host, port, basePath, apiStub);
+    public static TestServerClient createRecipeExecutor(Scheme scheme, String host, int port, String basePath, TestServerApi apiStub){
+        return new TestServerClient(scheme, host, port, basePath, apiStub);
     }
 }
